pi5-standalone/requirements.txt
pygame==2.6.1
pyyaml==6.0.3
flask==3.1.0
